feat: 重构 AI API，迁移和优化多个功能模块，包括聊天处理、MCP 调用检测、事件管理等
