+++ src-tauri/src/api/ai/mod.rs
@@ -1,0 +1,9 @@
+pub mod types;
+pub mod events;
+pub mod mcp;
+pub mod config;
+pub mod chat;
+pub mod conversation;
+pub mod title;
+
+
