+++ src-tauri/src/api/ai_api.rs
@@ -1,9 +1,5 @@
-use crate::api::assistant_api::{
-    get_assistant, get_assistant_mcp_servers_with_tools, MCPServerWithTools,
-};
+use crate::api::assistant_api::get_assistant;
 use crate::api::genai_client;
-use crate::api::mcp_execution_api::create_mcp_tool_call;
-use crate::db::assistant_db::AssistantModelConfig;
 use crate::db::conversation_db::{AttachmentType, Repository};
 use crate::db::conversation_db::{Conversation, ConversationDatabase, Message, MessageAttachment};
 use crate::db::llm_db::LLMDatabase;
@@ -14,1395 +10,68 @@ use crate::template_engine::TemplateEngine;
 use crate::{AppState, FeatureConfigState};
 use anyhow::Context;
 use anyhow::Error;
-use serde::{Deserialize, Serialize};
+// use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 use tauri::Emitter;
 use tauri::State;
 use tokio_util::sync::CancellationToken;
+use crate::api::ai::events::{ConversationEvent, MessageAddEvent, MessageUpdateEvent, ERROR_NOTIFICATION_EVENT};
+use crate::api::ai::types::{AiRequest, AiResponse};
 
-/// Conversation事件数据结构
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct ConversationEvent {
-    r#type: String,
-    data: serde_json::Value,
-}
+// 事件类型已抽到 crate::api::ai::events
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct MessageAddEvent {
-    message_id: i64,
-    message_type: String,
-}
+// use futures::StreamExt; // 已迁移到 ai::chat
+use genai::chat::ChatRequest;
+// use genai::Client; // 已迁移到 ai::chat
+// use tokio::time::{sleep, Duration}; // 已迁移到 ai::*
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct MessageUpdateEvent {
-    message_id: i64,
-    message_type: String,
-    content: String,
-    is_done: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct MessageTypeEndEvent {
-    message_id: i64,
-    message_type: String,
-    duration_ms: i64,
-    end_time: chrono::DateTime<chrono::Utc>,
-}
-
-use futures::StreamExt;
-use genai::chat::{ChatMessage, ChatOptions, ChatRequest, ContentPart};
-use genai::Client;
-use tokio::time::{sleep, Duration};
-
-/// 检测并处理消息中的 MCP 工具调用
-async fn detect_and_process_mcp_calls(
-    app_handle: &tauri::AppHandle,
-    conversation_id: i64,
-    message_id: i64,
-    content: &str,
-) -> Result<(), anyhow::Error> {
-    // 使用正则表达式匹配 MCP 工具调用
-    let mcp_regex = regex::Regex::new(r"<mcp_tool_call>\s*<server_name>([^<]*)</server_name>\s*<tool_name>([^<]*)</tool_name>\s*<parameters>([\s\S]*?)</parameters>\s*</mcp_tool_call>").unwrap();
-    
-    for cap in mcp_regex.captures_iter(content) {
-        let server_name = cap[1].trim().to_string();
-        let tool_name = cap[2].trim().to_string();
-        let parameters = cap[3].trim().to_string();
-        
-        println!(
-            "Detected MCP call: server={}, tool={}, params={}",
-            server_name, tool_name, parameters
-        );
-        
-        // 创建 MCP 工具调用记录
-        match create_mcp_tool_call(
-            app_handle.clone(),
-            conversation_id,
-            Some(message_id),
-            server_name,
-            tool_name,
-            parameters,
-        )
-        .await
-        {
-            Ok(tool_call) => {
-                println!("Created MCP tool call with ID: {}", tool_call.id);
-            }
-            Err(e) => {
-                eprintln!("Failed to create MCP tool call: {}", e);
-            }
-        }
-    }
-    
-    Ok(())
-}
+// 本地 MCP 工具调用检测逻辑已迁移至 crate::api::ai::mcp::detect_and_process_mcp_calls
 
 use super::assistant_api::AssistantDetail;
+use crate::api::ai::config::{ChatConfig, ConfigBuilder};
+use crate::api::ai::conversation::build_chat_messages;
+use crate::api::ai::mcp::{collect_mcp_info_for_assistant, format_mcp_prompt};
+use tokio::time::{sleep, Duration};
+use crate::api::ai::chat::{handle_non_stream_chat as ai_handle_non_stream_chat, handle_stream_chat as ai_handle_stream_chat};
+// 重复导入移除（已在文件顶部导入 AiRequest/AiResponse）
+use crate::api::ai::title::generate_title;
 
-// 事件名称常量
-const _MESSAGE_FINISH_EVENT: &str = "Tea::Event::MessageFinish";
-const TITLE_CHANGE_EVENT: &str = "title_change";
-const ERROR_NOTIFICATION_EVENT: &str = "conversation-window-error-notification";
 
-/// 重试配置
-const MAX_RETRY_ATTEMPTS: u32 = 1;
-const RETRY_DELAY_MS: u64 = 1000;
+// 事件名称常量已抽到 crate::api::ai::events
 
-/// AI聊天配置
-#[derive(Debug, Clone)]
-struct ChatConfig {
-    model_name: String,
-    stream: bool,
-    chat_options: ChatOptions,
-    client: Client,
-}
+// 重试配置常量已从 crate::api::ai::config 引入
 
-/// 配置构建器
-struct ConfigBuilder;
+// ChatConfig/ConfigBuilder 已迁移至 crate::api::ai::config
 
-impl ConfigBuilder {
-    /// 构建聊天选项
-    fn build_chat_options(config_map: &HashMap<String, String>) -> ChatOptions {
-        let mut chat_options = ChatOptions::default();
-
-        if let Some(temp_str) = config_map.get("temperature") {
-            if let Ok(temp) = temp_str.parse::<f64>() {
-                chat_options = chat_options.with_temperature(temp);
-            }
-        }
-
-        if let Some(max_tokens_str) = config_map.get("max_tokens") {
-            if let Ok(max_tokens) = max_tokens_str.parse::<u32>() {
-                chat_options = chat_options.with_max_tokens(max_tokens);
-            }
-        }
-
-        if let Some(top_p_str) = config_map.get("top_p") {
-            if let Ok(top_p) = top_p_str.parse::<f64>() {
-                chat_options = chat_options.with_top_p(top_p);
-            }
-        }
-
-        chat_options
-    }
-
-    /// 合并模型配置
-    fn merge_model_configs(
-        base_configs: Vec<AssistantModelConfig>,
-        model_detail: &crate::db::llm_db::ModelDetail,
-        override_configs: Option<HashMap<String, serde_json::Value>>,
-    ) -> Vec<AssistantModelConfig> {
-        let mut model_config_clone = base_configs;
-        model_config_clone.push(AssistantModelConfig {
-            id: 0,
-            assistant_id: model_detail.model.id, // 使用正确的字段
-            assistant_model_id: model_detail.model.id,
-            name: "model".to_string(),
-            value: Some(model_detail.model.code.clone()),
-            value_type: "string".to_string(),
-        });
-
-        if let Some(override_configs) = override_configs {
-            for (key, value) in override_configs {
-                let value_type = match &value {
-                    serde_json::Value::String(_) => "string",
-                    serde_json::Value::Number(_) => "number",
-                    serde_json::Value::Bool(_) => "boolean",
-                    serde_json::Value::Array(_) => "array",
-                    serde_json::Value::Object(_) => "object",
-                    serde_json::Value::Null => "null",
-                };
-
-                let value_str = match value {
-                    serde_json::Value::String(s) => s,
-                    other => other.to_string(),
-                };
-
-                // 查找是否已存在该配置
-                if let Some(existing_config) = model_config_clone.iter_mut().find(|c| c.name == key)
-                {
-                    existing_config.value = Some(value_str);
-                    existing_config.value_type = value_type.to_string();
-                } else {
-                    // 添加新配置
-                    model_config_clone.push(AssistantModelConfig {
-                        id: 0,
-                        assistant_id: model_detail.model.id,
-                        assistant_model_id: model_detail.model.id,
-                        name: key,
-                        value: Some(value_str),
-                        value_type: value_type.to_string(),
-                    });
-                }
-            }
-        }
-
-        model_config_clone
-    }
-}
+// ConfigBuilder 实现已迁移
 
 /// 将消息列表转换为 ChatMessage 格式，并处理多媒体附件
 /// 过滤掉推理类型的消息，tool_result消息参与AI对话历史但不显示在UI
-fn build_chat_messages(
-    init_message_list: &[(String, String, Vec<MessageAttachment>)],
-) -> Vec<ChatMessage> {
-    let mut chat_messages = Vec::new();
-
-    for (message_type, content, attachments) in init_message_list {
-        // 过滤掉推理类型的消息（不参与AI对话历史），但tool_result要参与对话历史
-        if message_type == "reasoning" {
-            continue;
-        }
-
-        // 将 response 类型转换为 assistant 角色，tool_result 也作为 assistant 角色（工具执行结果）
-        let role = if message_type == "response" || message_type == "tool_result" {
-            "assistant"
-        } else {
-            message_type.as_str()
-        };
-
-        // 如果没有附件，使用简单的文本消息
-        if attachments.is_empty() {
-            match role {
-                "system" => chat_messages.push(ChatMessage::system(content)),
-                "user" => chat_messages.push(ChatMessage::user(content)),
-                "assistant" => chat_messages.push(ChatMessage::assistant(content)),
-                _ => {}
-            }
-            continue;
-        }
-
-        // 如果有附件，使用 ContentPart 来构建消息
-        let mut content_parts = vec![ContentPart::from_text(content)];
-
-        // 处理各种类型的附件
-        for attachment in attachments {
-            match attachment.attachment_type {
-                crate::db::conversation_db::AttachmentType::Image => {
-                    // 图像附件
-                    if let Some(content) = &attachment.attachment_content {
-                        // 解析 data URL 格式的内容，提取 MIME type 和纯 base64 内容
-                        if let Some((content_type, base64_content)) = parse_data_url(content) {
-                            content_parts.push(ContentPart::from_image_base64(
-                                &content_type,
-                                &*base64_content,
-                            ));
-                        }
-                    } else if let Some(url) = &attachment.attachment_url {
-                        // 推断图像的媒体类型
-                        let media_type = infer_media_type_from_url(url);
-                        content_parts.push(ContentPart::from_image_url(&media_type, url.as_str()));
-                    }
-                }
-                crate::db::conversation_db::AttachmentType::Text => {
-                    // 文本附件
-                    if let Some(attachment_content) = &attachment.attachment_content {
-                        let file_name = attachment.attachment_url.as_deref().unwrap_or("未知文件");
-                        content_parts.push(ContentPart::from_text(&format!(
-                            "\n\n[文本附件: {}]\n{}",
-                            file_name, attachment_content
-                        )));
-                    }
-                }
-                crate::db::conversation_db::AttachmentType::PDF
-                | crate::db::conversation_db::AttachmentType::Word
-                | crate::db::conversation_db::AttachmentType::PowerPoint
-                | crate::db::conversation_db::AttachmentType::Excel => {
-                    // 其他文档类型，作为文本内容处理
-                    if let Some(attachment_content) = &attachment.attachment_content {
-                        let file_name = attachment.attachment_url.as_deref().unwrap_or("未知文档");
-                        let file_type = match attachment.attachment_type {
-                            crate::db::conversation_db::AttachmentType::PDF => "PDF文档",
-                            crate::db::conversation_db::AttachmentType::Word => "Word文档",
-                            crate::db::conversation_db::AttachmentType::PowerPoint => {
-                                "PowerPoint文档"
-                            }
-                            crate::db::conversation_db::AttachmentType::Excel => "Excel文档",
-                            _ => "文档",
-                        };
-                        content_parts.push(ContentPart::from_text(&format!(
-                            "\n\n[{}: {}]\n{}",
-                            file_type, file_name, attachment_content
-                        )));
-                    }
-                }
-            }
-        }
-
-        // 创建包含多个内容部分的消息
-        match role {
-            "system" => {
-                // 系统消息通常不支持多媒体内容，将所有内容合并为文本
-                let combined_text = content_parts
-                    .iter()
-                    .map(|part| {
-                        // 注意：这里假设 ContentPart 有某种方式提取文本，
-                        // 实际情况可能需要根据 genai 库的具体实现调整
-                        match part {
-                            // 这里需要根据实际的 ContentPart API 来实现
-                            _ => content.clone(), // 临时处理
-                        }
-                    })
-                    .collect::<Vec<_>>()
-                    .join("");
-                chat_messages.push(ChatMessage::system(&combined_text));
-            }
-            "user" => {
-                chat_messages.push(ChatMessage::user(content_parts));
-            }
-            "assistant" => {
-                // 助手消息也通常是纯文本，将内容合并
-                let combined_text = content_parts
-                    .iter()
-                    .map(|_| content.clone()) // 临时处理
-                    .collect::<Vec<_>>()
-                    .join("");
-                chat_messages.push(ChatMessage::assistant(&combined_text));
-            }
-            _ => {}
-        }
-    }
-    println!("================================ Chat Messages (Filtered) ===============================================");
-    println!("{:?}", chat_messages);
-    println!("================================ Chat Messages End ===============================================");
-    chat_messages
-}
+// 已迁移至 crate::api::ai::conversation::build_chat_messages
 
 /// 根据URL推断图像的媒体类型
-fn infer_media_type_from_url(url: &str) -> String {
-    let url_lower = url.to_lowercase();
-    if url_lower.ends_with(".jpg") || url_lower.ends_with(".jpeg") {
-        "image/jpeg".to_string()
-    } else if url_lower.ends_with(".png") {
-        "image/png".to_string()
-    } else if url_lower.ends_with(".gif") {
-        "image/gif".to_string()
-    } else if url_lower.ends_with(".webp") {
-        "image/webp".to_string()
-    } else if url_lower.ends_with(".bmp") {
-        "image/bmp".to_string()
-    } else {
-        "image/jpeg".to_string() // 默认值
-    }
-}
+// 已迁移至 crate::api::ai::conversation::infer_media_type_from_url
 
 /// 解析 data URL 格式的内容，提取 MIME type 和纯 base64 内容
 /// 支持格式：data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
-fn parse_data_url(data_url: &str) -> Option<(String, String)> {
-    if !data_url.starts_with("data:") {
-        return None;
-    }
-
-    let parts: Vec<&str> = data_url.splitn(2, ',').collect();
-    if parts.len() != 2 {
-        return None;
-    }
-
-    let header = parts[0];
-    let content = parts[1];
-
-    // 提取 MIME type
-    let header_without_data = header.strip_prefix("data:")?;
-    let mime_type = if let Some(semicolon_pos) = header_without_data.find(';') {
-        &header_without_data[..semicolon_pos]
-    } else {
-        header_without_data
-    };
-
-    // 检查是否包含 base64 标识
-    if !header.contains("base64") {
-        return None;
-    }
-
-    Some((mime_type.to_string(), content.to_string()))
-}
+// 已迁移至 crate::api::ai::conversation::parse_data_url
 
 /// 清理消息令牌
-async fn cleanup_token(
-    tokens: &Arc<tokio::sync::Mutex<HashMap<i64, CancellationToken>>>,
-    message_id: i64,
-) {
-    let mut map = tokens.lock().await;
-    map.remove(&message_id);
-}
+// 已迁移至 crate::api::ai::conversation::cleanup_token
 
 /// 处理消息类型结束事件
-async fn handle_message_type_end(
-    message_id: i64,
-    message_type: &str,
-    content: &str,
-    start_time: chrono::DateTime<chrono::Utc>,
-    conversation_db: &ConversationDatabase,
-    window: &tauri::Window,
-    conversation_id: i64,
-    app_handle: &tauri::AppHandle,
-) -> Result<(), anyhow::Error> {
-    let end_time = chrono::Utc::now();
-    let duration_ms = end_time.timestamp_millis() - start_time.timestamp_millis();
+// 已迁移至 crate::api::ai::conversation::handle_message_type_end
 
-    // 更新数据库的finish_time
-    conversation_db
-        .message_repo()?
-        .update_finish_time(message_id)?;
+// MCP 信息数据结构已抽到 crate::api::ai::mcp
 
-    // 如果是response消息，检测MCP工具调用
-    if message_type == "response" {
-        if let Err(e) = detect_and_process_mcp_calls(app_handle, conversation_id, message_id, content).await {
-            eprintln!("Failed to detect MCP calls: {}", e);
-        }
-    }
+// AiRequest/AiResponse 已抽到 crate::api::ai::types
 
-    // 发送类型结束事件
-    let type_end_event = ConversationEvent {
-        r#type: "message_type_end".to_string(),
-        data: serde_json::to_value(MessageTypeEndEvent {
-            message_id,
-            message_type: message_type.to_string(),
-            duration_ms,
-            end_time,
-        })
-        .unwrap(),
-    };
-    let _ = window.emit(
-        format!("conversation_event_{}", conversation_id).as_str(),
-        type_end_event,
-    );
+// 收集助手的 MCP 信息已迁移至 crate::api::ai::mcp::collect_mcp_info_for_assistant
 
-    // 发送最终的更新事件，标记为完成
-    let final_update_event = ConversationEvent {
-        r#type: "message_update".to_string(),
-        data: serde_json::to_value(MessageUpdateEvent {
-            message_id,
-            message_type: message_type.to_string(),
-            content: content.to_string(),
-            is_done: true,
-        })
-        .unwrap(),
-    };
-    let _ = window.emit(
-        format!("conversation_event_{}", conversation_id).as_str(),
-        final_update_event,
-    );
+// 格式化 MCP 提示词逻辑已迁移至 crate::api::ai::mcp::format_mcp_prompt
 
-    Ok(())
-}
 
-/// MCP 信息数据结构，用于存储助手的 MCP 配置信息
-#[derive(Debug, Clone)]
-struct MCPInfoForAssistant {
-    pub enabled_servers: Vec<MCPServerWithTools>,
-    pub use_native_toolcall: bool,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct AiRequest {
-    pub conversation_id: String,
-    pub assistant_id: i64,
-    pub prompt: String,
-    pub model: Option<String>,
-    pub temperature: Option<f32>,
-    pub top_p: Option<f32>,
-    pub max_tokens: Option<u32>,
-    pub stream: Option<bool>,
-    pub attachment_list: Option<Vec<i64>>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct AiResponse {
-    conversation_id: i64,
-    request_prompt_result_with_context: String,
-}
-
-/// 收集助手的 MCP 信息
-async fn collect_mcp_info_for_assistant(
-    app_handle: &tauri::AppHandle,
-    assistant_id: i64,
-) -> Result<MCPInfoForAssistant, AppError> {
-    // 1. 获取原生 toolcall 配置
-    let use_native_toolcall = match super::assistant_api::get_assistant_field_value(
-        app_handle.clone(),
-        assistant_id,
-        "use_native_toolcall",
-    ) {
-        Ok(value) => {
-            let result = value == "true";
-            result
-        }
-        Err(e) => {
-            println!(
-                "Failed to get native toolcall config: {}, using default (false)",
-                e
-            );
-            false
-        }
-    };
-
-    // 2. 直接使用已验证的 get_assistant_mcp_servers_with_tools 函数
-    let enabled_servers = get_assistant_mcp_servers_with_tools(app_handle.clone(), assistant_id)
-        .await
-        .map_err(|e| AppError::DatabaseError(format!("Failed to get MCP servers: {}", e)))?;
-
-    println!("Enabled servers: {:?}", enabled_servers);
-
-    Ok(MCPInfoForAssistant {
-        enabled_servers,
-        use_native_toolcall,
-    })
-}
-
-/// 格式化 MCP 提示词，添加 AI 约束和工具信息
-async fn format_mcp_prompt(
-    assistant_prompt_result: String,
-    mcp_info: &MCPInfoForAssistant,
-) -> String {
-    // MCP 约束提示词
-    let mcp_constraint_prompt = r#"
-# MCP (Model Context Protocol) 工具使用规范
-
-作为 AI 助手，你可以使用以下 MCP 工具来执行各种任务。请严格遵守以下规则：
-
-## 使用原则
-1. 你仅能调用以下提供的工具，不能够调用未提及的工具
-2. 优先使用最适合任务的工具
-3. 使用工具你可以获取到你需要的数据，不要怀疑工具的返回结果
-4. 每次只调用一个工具，等待之后返回的结果
-5. 工具调用放置在你回复的最后
-
-## 输出格式
-当需要调用 MCP 工具时，请使用以下 XML 格式：
-
-```xml
-<mcp_tool_call>
-<server_name>服务器名称</server_name>
-<tool_name>工具名称</tool_name>
-<parameters>
-{
-  "parameter1": "value1",
-  "parameter2": "value2"
-}
-</parameters>
-</mcp_tool_call>
-```
-
-## 重要注意事项
-- 参数必须是有效的 JSON 格式
-- 如果工具不需要参数，parameters 标签内应该为空对象 {}
-
-"#;
-
-    // 格式化可用的 MCP 服务器和工具信息
-    let mut tools_info = String::from("\n## 可用的 MCP 工具\n\n");
-
-    for server_details in &mcp_info.enabled_servers {
-        tools_info.push_str(&format!("### 服务器: {}\n", server_details.name));
-
-        tools_info.push_str("\n#### 可用工具:\n\n");
-
-        for tool in &server_details.tools {
-            tools_info.push_str(&format!("**{}** \n", tool.name));
-            tools_info.push_str(&format!(" - description: {}\n", tool.description));
-            tools_info.push_str(&format!(" - parameters: {}\n", tool.parameters));
-
-            tools_info.push_str("\n\n");
-        }
-
-        tools_info.push_str("\n---\n\n");
-    }
-
-    // 组合最终的提示词
-    format!(
-        "{}\n{}\n{}\n{}",
-        mcp_constraint_prompt, tools_info, "## 原始助手指令\n", assistant_prompt_result
-    )
-}
-
-/// 处理流式聊天
-async fn handle_stream_chat(
-    client: &Client,
-    model_name: &str,
-    chat_request: &ChatRequest,
-    chat_options: &ChatOptions,
-    conversation_id: i64,
-    cancel_token: &CancellationToken,
-    conversation_db: &ConversationDatabase,
-    tokens: &Arc<tokio::sync::Mutex<HashMap<i64, CancellationToken>>>,
-    window: &tauri::Window,
-    app_handle: &tauri::AppHandle,
-    need_generate_title: bool,
-    user_prompt: String,
-    config_feature_map: HashMap<String, HashMap<String, FeatureConfig>>,
-    generation_group_id_override: Option<String>, // 新增参数，用于regenerate时复用group_id
-    parent_group_id_override: Option<String>,     // 新增参数，用于设置parent_group_id
-    llm_model_id: i64,                            // 模型ID
-    llm_model_name: String,                       // 模型名称
-) -> Result<(), anyhow::Error> {
-    // 添加重试逻辑
-    let mut attempts: u32 = 0;
-    let app_handle_clone = app_handle.clone();
-    let chat_stream_result = loop {
-        match client
-            .exec_chat_stream(model_name, chat_request.clone(), Some(chat_options))
-            .await
-        {
-            Ok(response) => break Ok(response),
-            Err(e) => {
-                attempts += 1;
-                if attempts > MAX_RETRY_ATTEMPTS {
-                    eprintln!("Chat stream failed after {} attempts: {}", attempts, e);
-                    break Err(e);
-                }
-                eprintln!(
-                    "Chat stream attempt {} failed: {}, retrying...",
-                    attempts, e
-                );
-                sleep(Duration::from_millis(RETRY_DELAY_MS)).await;
-            }
-        }
-    };
-
-    match chat_stream_result {
-        Ok(chat_stream_response) => {
-            let mut chat_stream = chat_stream_response.stream;
-            let mut reasoning_content = String::new();
-            let mut response_content = String::new();
-            let mut reasoning_message_id: Option<i64> = None;
-            let mut response_message_id: Option<i64> = None;
-
-            // 为这次生成确定组ID：优先使用传入的group_id，否则创建新的
-            let generation_group_id =
-                generation_group_id_override.unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-
-            // 检查是否为重新生成操作（如果有parent_group_id_override则为重新生成）
-            let is_regeneration = parent_group_id_override.is_some();
-            let mut group_merge_event_emitted = false;
-
-            // 状态跟踪变量
-            let mut current_output_type: Option<String> = None;
-            let mut reasoning_start_time: Option<chrono::DateTime<chrono::Utc>> = None;
-            let mut response_start_time: Option<chrono::DateTime<chrono::Utc>> = None;
-
-            loop {
-                tokio::select! {
-                    stream_result = chat_stream.next() => {
-                        match stream_result {
-                            Some(Ok(stream_event)) => {
-                                use genai::chat::ChatStreamEvent;
-                                match stream_event {
-                                    ChatStreamEvent::Start => {
-                                        // 流开始，暂时不做处理
-                                    }
-                                    ChatStreamEvent::Chunk(chunk) => {
-                                        // 检查是否需要结束reasoning状态
-                                        if current_output_type == Some("reasoning".to_string()) {
-                                            if let (Some(msg_id), Some(start_time)) = (reasoning_message_id, reasoning_start_time) {
-                                                if let Err(e) = handle_message_type_end(
-                                                    msg_id,
-                                                    "reasoning",
-                                                    &reasoning_content,
-                                                    start_time,
-                                                    &conversation_db,
-                                                    &window,
-                                                    conversation_id,
-                                                    &app_handle_clone,
-                                                ).await {
-                                                    eprintln!("Failed to handle reasoning type end: {}", e);
-                                                }
-                                            }
-                                        }
-
-                                        // 切换到response状态
-                                        if current_output_type != Some("response".to_string()) {
-                                            current_output_type = Some("response".to_string());
-                                        }
-
-                                        // 处理正常回答内容
-                                        response_content.push_str(&chunk.content);
-
-                                        // 如果还没有创建 response 消息，创建一个
-                                        if response_message_id.is_none() {
-                                            let now = chrono::Utc::now();
-                                            response_start_time = Some(now);
-
-                                            let new_message = conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .create(&Message {
-                                                    id: 0,
-                                                    parent_id: None,
-                                                    conversation_id,
-                                                    message_type: "response".to_string(),
-                                                    content: response_content.clone(),
-                                                    llm_model_id: Some(llm_model_id),
-                                                    llm_model_name: Some(llm_model_name.clone()),
-                                                    created_time: now,
-                                                    start_time: Some(now),
-                                                    finish_time: None,
-                                                    token_count: 0,
-                                                    generation_group_id: Some(generation_group_id.clone()),
-                                                    parent_group_id: parent_group_id_override.clone(),
-                                                })
-                                                .unwrap();
-                                            response_message_id = Some(new_message.id);
-
-                                            // 发送消息添加事件
-                                            let add_event = ConversationEvent {
-                                                r#type: "message_add".to_string(),
-                                                data: serde_json::to_value(MessageAddEvent {
-                                                    message_id: new_message.id,
-                                                    message_type: "response".to_string(),
-                                                }).unwrap(),
-                                            };
-                                            let _ = window.emit(
-                                                format!("conversation_event_{}", conversation_id).as_str(),
-                                                add_event
-                                            );
-
-                                            // 如果是重新生成并且还没有发送组合并事件，则发送
-                                            if is_regeneration && !group_merge_event_emitted {
-                                                if let Some(ref parent_group_id) = parent_group_id_override {
-                                                    let group_merge_event = serde_json::json!({
-                                                        "type": "group_merge",
-                                                        "data": {
-                                                            "original_group_id": parent_group_id,
-                                                            "new_group_id": generation_group_id.clone(),
-                                                            "is_regeneration": true,
-                                                            "first_message_id": new_message.id,
-                                                            "conversation_id": conversation_id
-                                                        }
-                                                    });
-                                                    let _ = window.emit(
-                                                        format!("conversation_event_{}", conversation_id).as_str(),
-                                                        group_merge_event
-                                                    );
-                                                    group_merge_event_emitted = true;
-                                                }
-                                            }
-                                        }
-
-                                        // 更新消息内容
-                                        if let Some(msg_id) = response_message_id {
-                                            let mut message = conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .read(msg_id)
-                                                .unwrap()
-                                                .unwrap();
-                                            message.content = response_content.clone();
-                                            conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .update(&message)
-                                                .unwrap();
-
-                                            // 发送消息更新事件
-                                            let update_event = ConversationEvent {
-                                                r#type: "message_update".to_string(),
-                                                data: serde_json::to_value(MessageUpdateEvent {
-                                                    message_id: msg_id,
-                                                    message_type: "response".to_string(),
-                                                    content: response_content.clone(),
-                                                    is_done: false,
-                                                }).unwrap(),
-                                            };
-                                            let _ = window.emit(
-                                                format!("conversation_event_{}", conversation_id).as_str(),
-                                                update_event
-                                            );
-                                        }
-                                    }
-                                    ChatStreamEvent::ReasoningChunk(reasoning_chunk) => {
-                                        // 切换到reasoning状态
-                                        if current_output_type != Some("reasoning".to_string()) {
-                                            current_output_type = Some("reasoning".to_string());
-                                        }
-
-                                        // 处理推理内容
-                                        reasoning_content.push_str(&reasoning_chunk.content);
-
-                                        // 如果还没有创建 reasoning 消息，创建一个
-                                        if reasoning_message_id.is_none() {
-                                            let now = chrono::Utc::now();
-                                            reasoning_start_time = Some(now);
-
-                                            let new_message = conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .create(&Message {
-                                                    id: 0,
-                                                    parent_id: None,
-                                                    conversation_id,
-                                                    message_type: "reasoning".to_string(),
-                                                    content: reasoning_content.clone(),
-                                                    llm_model_id: Some(llm_model_id),
-                                                    llm_model_name: Some(llm_model_name.clone()),
-                                                    created_time: now,
-                                                    start_time: Some(now),
-                                                    finish_time: None,
-                                                    token_count: 0,
-                                                    generation_group_id: Some(generation_group_id.clone()),
-                                                    parent_group_id: parent_group_id_override.clone(),
-                                                })
-                                                .unwrap();
-                                            reasoning_message_id = Some(new_message.id);
-
-                                            // 发送消息添加事件
-                                            let add_event = ConversationEvent {
-                                                r#type: "message_add".to_string(),
-                                                data: serde_json::to_value(MessageAddEvent {
-                                                    message_id: new_message.id,
-                                                    message_type: "reasoning".to_string(),
-                                                }).unwrap(),
-                                            };
-                                            let _ = window.emit(
-                                                format!("conversation_event_{}", conversation_id).as_str(),
-                                                add_event
-                                            );
-
-                                            // 如果是重新生成并且还没有发送组合并事件，则发送
-                                            if is_regeneration && !group_merge_event_emitted {
-                                                if let Some(ref parent_group_id) = parent_group_id_override {
-                                                    let group_merge_event = serde_json::json!({
-                                                        "type": "group_merge",
-                                                        "data": {
-                                                            "original_group_id": parent_group_id,
-                                                            "new_group_id": generation_group_id.clone(),
-                                                            "is_regeneration": true,
-                                                            "first_message_id": new_message.id,
-                                                            "conversation_id": conversation_id
-                                                        }
-                                                    });
-                                                    let _ = window.emit(
-                                                        format!("conversation_event_{}", conversation_id).as_str(),
-                                                        group_merge_event
-                                                    );
-                                                    group_merge_event_emitted = true;
-                                                }
-                                            }
-                                        }
-
-                                        // 更新消息内容
-                                        if let Some(msg_id) = reasoning_message_id {
-                                            let mut message = conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .read(msg_id)
-                                                .unwrap()
-                                                .unwrap();
-                                            message.content = reasoning_content.clone();
-                                            conversation_db
-                                                .message_repo()
-                                                .unwrap()
-                                                .update(&message)
-                                                .unwrap();
-
-                                            // 发送消息更新事件
-                                            let update_event = ConversationEvent {
-                                                r#type: "message_update".to_string(),
-                                                data: serde_json::to_value(MessageUpdateEvent {
-                                                    message_id: msg_id,
-                                                    message_type: "reasoning".to_string(),
-                                                    content: reasoning_content.clone(),
-                                                    is_done: false,
-                                                }).unwrap(),
-                                            };
-                                            let _ = window.emit(
-                                                format!("conversation_event_{}", conversation_id).as_str(),
-                                                update_event
-                                            );
-                                        }
-                                    }
-                                    ChatStreamEvent::ToolCallChunk(_tool_chunk) => {
-                                        // 工具调用块，暂时忽略
-                                    }
-                                    ChatStreamEvent::End(_stream_end) => {
-                                        // 流结束，处理当前活跃的消息类型
-                                        match current_output_type.as_deref() {
-                                            Some("reasoning") => {
-                                                if let (Some(msg_id), Some(start_time)) = (reasoning_message_id, reasoning_start_time) {
-                                                    if let Err(e) = handle_message_type_end(
-                                                        msg_id,
-                                                        "reasoning",
-                                                        &reasoning_content,
-                                                        start_time,
-                                                        &conversation_db,
-                                                        &window,
-                                                        conversation_id,
-                                                        &app_handle_clone,
-                                                    ).await {
-                                                        eprintln!("Failed to handle reasoning type end: {}", e);
-                                                    }
-                                                }
-                                            }
-                                            Some("response") => {
-                                                if let (Some(msg_id), Some(start_time)) = (response_message_id, response_start_time) {
-                                                    if let Err(e) = handle_message_type_end(
-                                                        msg_id,
-                                                        "response",
-                                                        &response_content,
-                                                        start_time,
-                                                        &conversation_db,
-                                                        &window,
-                                                        conversation_id,
-                                                        &app_handle_clone,
-                                                    ).await {
-                                                        eprintln!("Failed to handle response type end: {}", e);
-                                                    }
-
-                                                    // 在 response 完成后自动生成标题
-                                                    if need_generate_title && !response_content.is_empty() {
-                                                        let app_handle_clone = app_handle.clone();
-                                                        let user_prompt_clone = user_prompt.clone();
-                                                        let response_content_clone = response_content.clone();
-                                                        let config_feature_map_clone = config_feature_map.clone();
-                                                        let window_clone = window.clone();
-
-                                                        tokio::spawn(async move {
-                                                            if let Err(e) = generate_title(
-                                                                &app_handle_clone,
-                                                                conversation_id,
-                                                                user_prompt_clone,
-                                                                response_content_clone,
-                                                                config_feature_map_clone,
-                                                                window_clone,
-                                                            ).await {
-                                                                eprintln!("Failed to generate title: {}", e);
-                                                            }
-                                                        });
-                                                    }
-                                                }
-                                            }
-                                            _ => {
-                                                // 没有活跃的消息类型，使用原有逻辑
-                                                return finish_stream_messages(
-                                                    &conversation_db,
-                                                    reasoning_message_id,
-                                                    response_message_id,
-                                                    &reasoning_content,
-                                                    &response_content,
-                                                    &window,
-                                                    conversation_id,
-                                                );
-                                            }
-                                        }
-                                        return Ok(());
-                                    }
-                                }
-                            },
-                            Some(Err(e)) => {
-                                eprintln!("Stream error: {}", e);
-                                cleanup_token(tokens, conversation_id).await;
-                                let err_msg = format!("Chat stream error: {}", e);
-
-                                // 发送直接错误通知事件到前端
-                                let _ = window.emit(ERROR_NOTIFICATION_EVENT, format!("Stream error: {}", e));
-
-                                // 发送错误事件到任一存在的消息
-                                if let Some(msg_id) = response_message_id.or(reasoning_message_id) {
-                                    let error_event = ConversationEvent {
-                                        r#type: "message_update".to_string(),
-                                        data: serde_json::to_value(MessageUpdateEvent {
-                                            message_id: msg_id,
-                                            message_type: "error".to_string(),
-                                            content: err_msg.clone(),
-                                            is_done: true,
-                                        }).unwrap(),
-                                    };
-                                    let _ = window.emit(
-                                        format!("conversation_event_{}", conversation_id).as_str(),
-                                        error_event
-                                    );
-                                } else {
-                                    // 如果没有现有消息，创建一个新的错误消息
-                                    let now = chrono::Utc::now();
-                                    let error_message_result = conversation_db
-                                        .message_repo()
-                                        .unwrap()
-                                        .create(&Message {
-                                            id: 0,
-                                            parent_id: None,
-                                            conversation_id,
-                                            message_type: "error".to_string(),
-                                            content: err_msg.clone(),
-                                            llm_model_id: Some(llm_model_id),
-                                            llm_model_name: Some(llm_model_name.clone()),
-                                            created_time: now,
-                                            start_time: Some(now),
-                                            finish_time: Some(now),
-                                            token_count: 0,
-                                            generation_group_id: Some(generation_group_id.clone()),
-                                            parent_group_id: parent_group_id_override.clone(),
-                                        });
-
-                                    if let Ok(error_message) = error_message_result {
-                                        // 发送消息添加事件
-                                        let add_event = ConversationEvent {
-                                            r#type: "message_add".to_string(),
-                                            data: serde_json::to_value(MessageAddEvent {
-                                                message_id: error_message.id,
-                                                message_type: "error".to_string(),
-                                            }).unwrap(),
-                                        };
-                                        let _ = window.emit(
-                                            format!("conversation_event_{}", conversation_id).as_str(),
-                                            add_event
-                                        );
-
-                                        // 发送消息更新事件（完成状态）
-                                        let update_event = ConversationEvent {
-                                            r#type: "message_update".to_string(),
-                                            data: serde_json::to_value(MessageUpdateEvent {
-                                                message_id: error_message.id,
-                                                message_type: "error".to_string(),
-                                                content: err_msg.clone(),
-                                                is_done: true,
-                                            }).unwrap(),
-                                        };
-                                        let _ = window.emit(
-                                            format!("conversation_event_{}", conversation_id).as_str(),
-                                            update_event
-                                        );
-                                    }
-                                }
-
-                                return Err(anyhow::anyhow!("Stream error: {}", e));
-                            },
-                            None => {
-                                // 流结束，处理当前活跃的消息类型
-                                match current_output_type.as_deref() {
-                                    Some("reasoning") => {
-                                        if let (Some(msg_id), Some(start_time)) = (reasoning_message_id, reasoning_start_time) {
-                                            if let Err(e) = handle_message_type_end(
-                                                msg_id,
-                                                "reasoning",
-                                                &reasoning_content,
-                                                start_time,
-                                                &conversation_db,
-                                                &window,
-                                                conversation_id,
-                                                &app_handle_clone,
-                                            ).await {
-                                                eprintln!("Failed to handle reasoning type end: {}", e);
-                                            }
-                                        }
-                                    }
-                                    Some("response") => {
-                                        if let (Some(msg_id), Some(start_time)) = (response_message_id, response_start_time) {
-                                            if let Err(e) = handle_message_type_end(
-                                                msg_id,
-                                                "response",
-                                                &response_content,
-                                                start_time,
-                                                &conversation_db,
-                                                &window,
-                                                conversation_id,
-                                                &app_handle_clone,
-                                            ).await {
-                                                eprintln!("Failed to handle response type end: {}", e);
-                                            }
-
-                                            // 在 response 完成后自动生成标题
-                                            if need_generate_title && !response_content.is_empty() {
-                                                let app_handle_clone = app_handle.clone();
-                                                let user_prompt_clone = user_prompt.clone();
-                                                let response_content_clone = response_content.clone();
-                                                let config_feature_map_clone = config_feature_map.clone();
-                                                let window_clone = window.clone();
-
-                                                tokio::spawn(async move {
-                                                    if let Err(e) = generate_title(
-                                                        &app_handle_clone,
-                                                        conversation_id,
-                                                        user_prompt_clone,
-                                                        response_content_clone,
-                                                        config_feature_map_clone,
-                                                        window_clone,
-                                                    ).await {
-                                                        eprintln!("Failed to generate title: {}", e);
-                                                    }
-                                                });
-                                            }
-                                        }
-                                    }
-                                    _ => {
-                                        // 没有活跃的消息类型，使用原有逻辑
-                                        return finish_stream_messages(
-                                            &conversation_db,
-                                            reasoning_message_id,
-                                            response_message_id,
-                                            &reasoning_content,
-                                            &response_content,
-                                            &window,
-                                            conversation_id,
-                                        );
-                                    }
-                                }
-                                return Ok(());
-                            }
-                        }
-                    }
-                    _ = cancel_token.cancelled() => {
-                        println!("Chat stream cancelled");
-                        return Ok(());
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            cleanup_token(tokens, conversation_id).await;
-            let err_msg = format!("Chat stream error: {}", e);
-
-            // 发送直接错误通知事件到前端
-            let _ = window.emit(
-                ERROR_NOTIFICATION_EVENT,
-                format!("Stream initialization error: {}", e),
-            );
-
-            // 为这次生成确定组ID：优先使用传入的group_id，否则创建新的
-            let generation_group_id =
-                generation_group_id_override.unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-            let now = chrono::Utc::now();
-
-            let error_message = conversation_db
-                .message_repo()
-                .unwrap()
-                .create(&Message {
-                    id: 0,
-                    parent_id: None,
-                    conversation_id,
-                    message_type: "error".to_string(),
-                    content: err_msg.clone(),
-                    llm_model_id: Some(llm_model_id),
-                    llm_model_name: Some(llm_model_name.clone()),
-                    created_time: now,
-                    start_time: Some(now),
-                    finish_time: Some(now), // Mark as finished immediately for errors
-                    token_count: 0,
-                    generation_group_id: Some(generation_group_id.clone()),
-                    parent_group_id: parent_group_id_override.clone(),
-                })
-                .unwrap();
-
-            // 发送错误事件
-            let error_event = ConversationEvent {
-                r#type: "message_add".to_string(),
-                data: serde_json::to_value(MessageAddEvent {
-                    message_id: error_message.id,
-                    message_type: "error".to_string(),
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                error_event,
-            );
-
-            // 发送消息更新事件（完成状态）
-            let update_event = ConversationEvent {
-                r#type: "message_update".to_string(),
-                data: serde_json::to_value(MessageUpdateEvent {
-                    message_id: error_message.id,
-                    message_type: "error".to_string(),
-                    content: err_msg.clone(),
-                    is_done: true,
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                update_event,
-            );
-
-            eprintln!("Chat stream error: {}", e);
-            return Err(anyhow::anyhow!("Chat stream error: {}", e));
-        }
-    }
-}
-
-/// 处理非流式聊天
-async fn handle_non_stream_chat(
-    client: &Client,
-    model_name: &str,
-    chat_request: &ChatRequest,
-    chat_options: &ChatOptions,
-    conversation_id: i64,
-    cancel_token: &CancellationToken,
-    conversation_db: &ConversationDatabase,
-    tokens: &Arc<tokio::sync::Mutex<HashMap<i64, CancellationToken>>>,
-    window: &tauri::Window,
-    app_handle: &tauri::AppHandle,
-    need_generate_title: bool,
-    user_prompt: String,
-    config_feature_map: HashMap<String, HashMap<String, FeatureConfig>>,
-    generation_group_id_override: Option<String>, // 新增参数，用于regenerate时复用group_id
-    parent_group_id_override: Option<String>,     // 新增参数，用于设置parent_group_id
-    llm_model_id: i64,                            // 模型ID
-    llm_model_name: String,                       // 模型名称
-) -> Result<(), anyhow::Error> {
-    // 为这次生成确定组ID：优先使用传入的group_id，否则创建新的
-    let generation_group_id = generation_group_id_override
-        .clone()
-        .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-
-    // 创建一个 response 类型的消息
-    let response_message = conversation_db
-        .message_repo()
-        .unwrap()
-        .create(&Message {
-            id: 0,
-            parent_id: None,
-            conversation_id,
-            message_type: "response".to_string(),
-            content: String::new(),
-            llm_model_id: Some(llm_model_id),
-            llm_model_name: Some(llm_model_name.clone()),
-            created_time: chrono::Utc::now(),
-            start_time: Some(chrono::Utc::now()),
-            finish_time: None,
-            token_count: 0,
-            generation_group_id: Some(generation_group_id),
-            parent_group_id: parent_group_id_override.clone(),
-        })
-        .unwrap();
-    let response_message_id = response_message.id;
-
-    // 发送消息添加事件
-    let add_event = ConversationEvent {
-        r#type: "message_add".to_string(),
-        data: serde_json::to_value(MessageAddEvent {
-            message_id: response_message_id,
-            message_type: "response".to_string(),
-        })
-        .unwrap(),
-    };
-    let _ = window.emit(
-        format!("conversation_event_{}", conversation_id).as_str(),
-        add_event,
-    );
-
-    let chat_result = tokio::select! {
-        result = async {
-            // 添加重试逻辑
-            let mut attempts = 0;
-            loop {
-                match client.exec_chat(model_name, chat_request.clone(), Some(chat_options)).await {
-                    Ok(response) => break Ok(response),
-                    Err(e) => {
-                        attempts += 1;
-                        if attempts > MAX_RETRY_ATTEMPTS {
-                            eprintln!("Chat request failed after {} attempts: {}", attempts, e);
-                            break Err(e);
-                        }
-                        eprintln!("Chat request attempt {} failed: {}, retrying...", attempts, e);
-                        sleep(Duration::from_millis(RETRY_DELAY_MS)).await;
-                    }
-                }
-            }
-        } => result,
-        _ = cancel_token.cancelled() => {
-            cleanup_token(tokens, conversation_id).await;
-            return Err(anyhow::anyhow!("Request cancelled"));
-        }
-    };
-
-    match chat_result {
-        Ok(chat_response) => {
-            let content = chat_response.first_text().unwrap_or("").to_string();
-            println!("Chat content: {}", content.clone());
-
-            // 更新消息内容
-            let mut message = conversation_db
-                .message_repo()
-                .unwrap()
-                .read(response_message_id)
-                .unwrap()
-                .unwrap();
-            message.content = content.clone();
-            conversation_db
-                .message_repo()
-                .unwrap()
-                .update(&message)
-                .unwrap();
-
-            conversation_db
-                .message_repo()
-                .unwrap()
-                .update_finish_time(response_message_id)
-                .unwrap();
-
-            // 发送消息更新事件（完成状态）
-            let update_event = ConversationEvent {
-                r#type: "message_update".to_string(),
-                data: serde_json::to_value(MessageUpdateEvent {
-                    message_id: response_message_id,
-                    message_type: "response".to_string(),
-                    content: content.clone(),
-                    is_done: true,
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                update_event,
-            );
-
-            // 在非流式消息完成后自动生成标题
-            if need_generate_title && !content.is_empty() {
-                let app_handle_clone = app_handle.clone();
-                let user_prompt_clone = user_prompt.clone();
-                let content_clone = content.clone();
-                let config_feature_map_clone = config_feature_map.clone();
-                let window_clone = window.clone();
-
-                tokio::spawn(async move {
-                    if let Err(e) = generate_title(
-                        &app_handle_clone,
-                        conversation_id,
-                        user_prompt_clone,
-                        content_clone,
-                        config_feature_map_clone,
-                        window_clone,
-                    )
-                    .await
-                    {
-                        eprintln!("Failed to generate title: {}", e);
-                    }
-                });
-            }
-
-            Ok(())
-        }
-        Err(e) => {
-            cleanup_token(tokens, conversation_id).await;
-            let err_msg = format!("Chat error: {}", e);
-            let now = chrono::Utc::now();
-
-            // 发送直接错误通知事件到前端
-            let _ = window.emit(
-                ERROR_NOTIFICATION_EVENT,
-                format!("Chat request error: {}", e),
-            );
-
-            // 为这次生成确定组ID：优先使用传入的group_id，否则创建新的
-            let generation_group_id = generation_group_id_override
-                .clone()
-                .unwrap_or_else(|| uuid::Uuid::new_v4().to_string());
-
-            let error_message = conversation_db
-                .message_repo()
-                .unwrap()
-                .create(&Message {
-                    id: 0,
-                    parent_id: None,
-                    conversation_id,
-                    message_type: "error".to_string(),
-                    content: err_msg.clone(),
-                    llm_model_id: Some(llm_model_id),
-                    llm_model_name: Some(llm_model_name.clone()),
-                    created_time: now,
-                    start_time: Some(now),
-                    finish_time: Some(now), // Mark as finished immediately for errors
-                    token_count: 0,
-                    generation_group_id: Some(generation_group_id.clone()),
-                    parent_group_id: parent_group_id_override.clone(),
-                })
-                .unwrap();
-
-            // 发送错误事件
-            let error_event = ConversationEvent {
-                r#type: "message_add".to_string(),
-                data: serde_json::to_value(MessageAddEvent {
-                    message_id: error_message.id,
-                    message_type: "error".to_string(),
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                error_event,
-            );
-
-            // 发送消息更新事件（完成状态）
-            let update_event = ConversationEvent {
-                r#type: "message_update".to_string(),
-                data: serde_json::to_value(MessageUpdateEvent {
-                    message_id: error_message.id,
-                    message_type: "error".to_string(),
-                    content: err_msg.clone(),
-                    is_done: true,
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                update_event,
-            );
-
-            eprintln!("Chat error: {}", e);
-            Err(anyhow::anyhow!("Chat error: {}", e))
-        }
-    }
-}
 
 #[tauri::command]
 pub async fn ask_ai(
@@ -1578,7 +247,7 @@ pub async fn ask_ai(
 
         if chat_config.stream {
             // 使用 genai 流式处理
-            handle_stream_chat(
+            ai_handle_stream_chat(
                 &chat_config.client,
                 &chat_config.model_name,
                 &chat_request,
@@ -1600,7 +269,7 @@ pub async fn ask_ai(
             .await?;
         } else {
             // Use genai non-streaming
-            handle_non_stream_chat(
+            ai_handle_non_stream_chat(
                 &chat_config.client,
                 &chat_config.model_name,
                 &chat_request,
@@ -1642,6 +311,9 @@ pub async fn cancel_ai(
     Ok(())
 }
 
+// 已迁移至 crate::api::ai::conversation::init_conversation（原地保留：调用路径较多）
+// 本地实现重命名以避免与导入的函数同名
+#[allow(dead_code)]
 fn init_conversation(
     app_handle: &tauri::AppHandle,
     assistant_id: i64,
@@ -1931,7 +603,7 @@ pub async fn regenerate_ai(
 
         if chat_config.stream {
             // 使用 genai 流式处理
-            handle_stream_chat(
+            ai_handle_stream_chat(
                 &chat_config.client,
                 &chat_config.model_name,
                 &chat_request,
@@ -1953,7 +625,7 @@ pub async fn regenerate_ai(
             .await?;
         } else {
             // Use genai non-streaming
-            handle_non_stream_chat(
+            ai_handle_non_stream_chat(
                 &chat_config.client,
                 &chat_config.model_name,
                 &chat_request,
@@ -2224,179 +896,7 @@ async fn initialize_conversation(
     ))
 }
 
-async fn generate_title(
-    app_handle: &tauri::AppHandle,
-    conversation_id: i64,
-    user_prompt: String,
-    content: String,
-    config_feature_map: HashMap<String, HashMap<String, FeatureConfig>>,
-    window: tauri::Window,
-) -> Result<(), AppError> {
-    // TODO 要检查下是否配置了对应的
-    let feature_config = config_feature_map.get("conversation_summary");
-    if let Some(config) = feature_config {
-        // model_id, prompt, summary_length
-        let provider_id = config
-            .get("provider_id")
-            .ok_or(AppError::NoConfigError("provider_id".to_string()))?
-            .value
-            .parse::<i64>()?;
-        let model_code = config
-            .get("model_code")
-            .ok_or(AppError::NoConfigError("model_code".to_string()))?
-            .value
-            .clone();
-        let prompt = config.get("prompt").unwrap().value.clone();
-        let summary_length = config
-            .get("summary_length")
-            .unwrap()
-            .value
-            .clone()
-            .parse::<i32>()
-            .unwrap();
-        let mut context = String::new();
-
-        if summary_length == -1 {
-            if content.is_empty() {
-                // 只有用户消息，没有助手回答
-                context.push_str(
-                    format!(
-                        "# user\n {} \n\n请为上述用户问题生成一个简洁的标题，不需要包含标点符号",
-                        user_prompt
-                    )
-                    .as_str(),
-                );
-            } else {
-                // 有用户消息和助手回答
-                context.push_str(
-                    format!(
-                        "# user\n {} \n\n#assistant\n {} \n\n请总结上述对话为标题，不需要包含标点符号",
-                        user_prompt, content
-                    )
-                    .as_str(),
-                );
-            }
-        } else {
-            let unsize_summary_length: usize = summary_length.try_into().unwrap();
-            if content.is_empty() {
-                // 只有用户消息，没有助手回答
-                if user_prompt.len() > unsize_summary_length {
-                    context.push_str(
-                        format!(
-                            "# user\n {} \n\n请为上述用户问题生成一个简洁的标题，不需要包含标点符号",
-                            user_prompt
-                                .chars()
-                                .take(unsize_summary_length)
-                                .collect::<String>()
-                        )
-                        .as_str(),
-                    );
-                } else {
-                    context.push_str(
-                        format!(
-                            "# user\n {} \n\n请为上述用户问题生成一个简洁的标题，不需要包含标点符号",
-                            user_prompt
-                        )
-                        .as_str(),
-                    );
-                }
-            } else {
-                // 有用户消息和助手回答
-                if user_prompt.len() > unsize_summary_length {
-                    context.push_str(
-                        format!(
-                            "# user\n {} \n\n请总结上述对话为标题，不需要包含标点符号",
-                            user_prompt
-                                .chars()
-                                .take(unsize_summary_length)
-                                .collect::<String>()
-                        )
-                        .as_str(),
-                    );
-                } else {
-                    let assistant_summary_length = unsize_summary_length - user_prompt.len();
-                    if content.len() > assistant_summary_length {
-                        context.push_str(format!("# user\n {} \n\n#assistant\n {} \n\n请总结上述对话为标题，不需要包含标点符号", user_prompt, content.chars().take(assistant_summary_length).collect::<String>()).as_str());
-                    } else {
-                        context.push_str(format!("# user\n {} \n\n#assistant\n {} \n\n请总结上述对话为标题，不需要包含标点符号", user_prompt, content).as_str());
-                    }
-                }
-            }
-        }
-
-        // 直接创建数据库连接
-        let llm_db = LLMDatabase::new(app_handle).map_err(AppError::from)?;
-        let model_detail = llm_db
-            .get_llm_model_detail(&provider_id, &model_code)
-            .unwrap();
-
-        // Create genai client with custom config
-        let client = genai_client::create_client_with_config(
-            &model_detail.configs,
-            &model_detail.model.code,
-            &model_detail.provider.api_type,
-        )?;
-
-        // Convert messages to ChatMessage format
-        let chat_messages = vec![ChatMessage::system(&prompt), ChatMessage::user(&context)];
-        let chat_request = ChatRequest::new(chat_messages);
-
-        // Use model code as model name
-        let model_name = &model_detail.model.code;
-
-        // 添加重试逻辑
-        let mut attempts = 0;
-        let response = loop {
-            match client
-                .exec_chat(model_name, chat_request.clone(), None)
-                .await
-            {
-                Ok(chat_response) => {
-                    break Ok(chat_response.first_text().unwrap_or("").to_string())
-                }
-                Err(e) => {
-                    attempts += 1;
-                    if attempts > MAX_RETRY_ATTEMPTS {
-                        eprintln!("Title generation failed after {} attempts: {}", attempts, e);
-                        break Err(e.to_string());
-                    }
-                    eprintln!(
-                        "Title generation attempt {} failed: {}, retrying...",
-                        attempts, e
-                    );
-                    sleep(Duration::from_millis(RETRY_DELAY_MS)).await;
-                }
-            }
-        };
-        match response {
-            Err(e) => {
-                println!("Chat error: {}", e);
-                let _ = window.emit(ERROR_NOTIFICATION_EVENT, "生成对话标题失败，请检查配置");
-            }
-            Ok(response_text) => {
-                println!("Chat content: {}", response_text.clone());
-
-                let conversation_db =
-                    ConversationDatabase::new(app_handle).map_err(AppError::from)?;
-                let _ = conversation_db
-                    .conversation_repo()
-                    .unwrap()
-                    .update_name(&Conversation {
-                        id: conversation_id,
-                        name: response_text.clone(),
-                        assistant_id: None,
-                        created_time: chrono::Utc::now(),
-                    });
-                window
-                    .emit(TITLE_CHANGE_EVENT, (conversation_id, response_text.clone()))
-                    .map_err(|e| e.to_string())
-                    .unwrap();
-            }
-        }
-    }
-
-    Ok(())
-}
+// 已迁移至 crate::api::ai::title::generate_title
 
 /// 重新生成对话标题
 #[tauri::command]
@@ -2452,65 +952,4 @@ pub async fn regenerate_conversation_title(
     Ok(())
 }
 
-/// 完成流式消息处理的统一函数
-fn finish_stream_messages(
-    conversation_db: &ConversationDatabase,
-    reasoning_message_id: Option<i64>,
-    response_message_id: Option<i64>,
-    reasoning_content: &str,
-    response_content: &str,
-    window: &tauri::Window,
-    conversation_id: i64,
-) -> Result<(), Error> {
-    // 如果只有 reasoning 没有 response，则结束 reasoning
-    if let Some(msg_id) = reasoning_message_id {
-        if response_message_id.is_none() {
-            conversation_db
-                .message_repo()
-                .unwrap()
-                .update_finish_time(msg_id)
-                .unwrap();
-
-            let complete_event = ConversationEvent {
-                r#type: "message_update".to_string(),
-                data: serde_json::to_value(MessageUpdateEvent {
-                    message_id: msg_id,
-                    message_type: "reasoning".to_string(),
-                    content: reasoning_content.to_string(),
-                    is_done: true,
-                })
-                .unwrap(),
-            };
-            let _ = window.emit(
-                format!("conversation_event_{}", conversation_id).as_str(),
-                complete_event,
-            );
-        }
-    }
-
-    // 结束 response 消息
-    if let Some(msg_id) = response_message_id {
-        conversation_db
-            .message_repo()
-            .unwrap()
-            .update_finish_time(msg_id)
-            .unwrap();
-
-        let complete_event = ConversationEvent {
-            r#type: "message_update".to_string(),
-            data: serde_json::to_value(MessageUpdateEvent {
-                message_id: msg_id,
-                message_type: "response".to_string(),
-                content: response_content.to_string(),
-                is_done: true,
-            })
-            .unwrap(),
-        };
-        let _ = window.emit(
-            format!("conversation_event_{}", conversation_id).as_str(),
-            complete_event,
-        );
-    }
-
-    Ok(())
-}
+// 完成流式消息处理的统一函数已迁移至 crate::api::ai::conversation::finish_stream_messages
