+++ src-tauri/src/api/mod.rs
@@ -3,6 +3,7 @@ pub mod artifacts_api;
 pub mod assistant_api;
 pub mod attachment_api;
 pub mod conversation_api;
+pub mod ai;
 pub mod genai_client;
 pub mod llm_api;
 pub mod mcp_api;
